refactor: 방송 스케줄 api 경로 수정 (변호사/의뢰인 나누기) - 변호사: /api/lawyer/schedule - 의뢰인: /api/client/schedule
