+++ src/main/java/com/lawnroad/broadcast/live/controller/CategoryController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.List;
 
 @RestController
-@RequestMapping("/api/category")
+@RequestMapping("/api/public/category")
 @RequiredArgsConstructor
 public class CategoryController {
